Fixed variants and webscrape
I had made the ExtractVariants file wrong, so there were some non-assigned characters in it, which are now removed. Because of this, they also needed to be removed from the webscrape, hence the FixScrape.py being added. I also decided that I should add all kangxi radicals as well as their supplements to the variants.txt, and as such, some more scraping had to be done (around 280 characters).